Moved get_accuracy to predict; swapped to sklearn accuracy metrics  - Moved accuracy calculations to predict  - Updated accuracy scores to use sklearn metrics
